fix parser

+++ pharser.h
@@ -25,7 +25,7 @@ typedef struct {
     char* table_name;
     int key;
     char* value;
-    unsigned long timestamp;
+    uint16_t timestamp;
 } package_insert;
 
 //DESCRIBE [NOMBRE_TABLA]?
